--- src/components/Post/PostComment.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const PostComment = () => {
-  return <div>PostComment</div>;
-};
-
-export default PostComment;
