+++ src/pages/PostPage/PostDetail.jsx
@@ -4,10 +4,9 @@ import Comment from '../../pages/ChatPage/Comment';
 import { AuthContextStore } from '../../context/AuthContext';
 import styled from 'styled-components';
 import Wrapper from '../../components/common/Wrapper/Wrapper';
-import { Link, useParams } from 'react-router-dom';
+import { useParams } from 'react-router-dom';
 import PostItem from '../../components/Post/PostItem';
-import PostUserProfileImg from '../../components/Post/PostUserProfileImg';
-import ModalButtonImg from '../../assets/icon/icon-more-vertical.png';
+import PostCommentItem from '../../components/Post/PostCommentItem';
 
 const PostDetail = () => {
   const [userInfo, setUserInfo] = useState({});
@@ -87,28 +86,7 @@ const PostDetail = () => {
         <CommentWrapper>
           <PostSection>
             <h2 className='a11y-hidden'>댓글 목록</h2>
-            <Ul>
-              {comments &&
-                comments.map((item) => {
-                  return (
-                    <Li key={item.id}>
-                      <UserAuth>
-                        <ProfileLink to={`/profile/${item.author.accountname}`}>
-                          <PostUserProfileImg size={'36px'} userProfileImg={item.author.image} />
-                        </ProfileLink>
-                        <NameLink to={`/profile/${item.author.accountname}`}>
-                          <Username>{item.author.username}</Username>
-                        </NameLink>
-                        <UploadTime>2minutes ago</UploadTime>
-                        <ButtonIcon>
-                          <img src={ModalButtonImg} alt='숨겨진 모달창 나타내기' />
-                        </ButtonIcon>
-                      </UserAuth>
-                      <PostComment>{item.content}</PostComment>
-                    </Li>
-                  );
-                })}
-            </Ul>
+            <PostCommentItem comments={comments} />
           </PostSection>
         </CommentWrapper>
       </Main>
@@ -140,44 +118,3 @@ const CommentWrapper = styled(Wrapper)`
 `;
 const PostArticle = styled.article``;
 const PostSection = styled.section``;
-const Ul = styled.ul``;
-const Li = styled.li`
-  position: relative;
-  & + & {
-    margin-top: 40px;
-  }
-`;
-const UserAuth = styled.div`
-  display: flex;
-`;
-const ProfileLink = styled(Link)``;
-const NameLink = styled(Link)`
-  margin: 4px 5px 0 0;
-`;
-const UploadTime = styled.span`
-  font-size: 10px;
-  line-height: 13px;
-  margin-top: 7.5px;
-  color: ${({ theme }) => theme.colors.txtColor};
-  &::before {
-    content: '·';
-    margin-right: 4px;
-  }
-`;
-const Username = styled.strong`
-  font-weight: 500;
-  /* font-size: 14px; */
-  line-height: 18px;
-`;
-const PostComment = styled.p`
-  /* font-size: 14px; */
-  padding-left: 48px;
-`;
-// 모달 버튼 아이콘
-const ButtonIcon = styled.button`
-  width: 18px;
-  position: absolute;
-  gap: 12px;
-  top: 5px;
-  right: 0;
-`;
